V 1.0.1 完成《Local Color Correction using Non-Linear Masking》算法，实际效果并不理想，出现过曝，全局提亮等问题，后续优化思路，图片分割，局部计算，可以根据图像直方图数据针对修改gamma矫正公式以适配人眼

+++ src/color_correction/color_correction.c
@@ -117,7 +117,7 @@ void LocalColorCorrection(unsigned char *Input, unsigned char *Output, int Width
     for (int mask = 0; mask < 256; mask++) {
         unsigned char *pLocalLut = LocalLut + (mask << 8);
         for (int pix = 0; pix < 256; pix++) {
-            pLocalLut[pix] = ClampToByte(255.0f * powf(pix / 255.0f, powf(2.0f, (128.0f - (255.0f - mask)) / 128.0f)));
+            pLocalLut[pix] = ClampToByte((int)(255.0f * powf((float)pix / 255.0f, powf(2.0f, (128.0f - (255.0f - mask)) / 128.0f))));
         }
     }
     Grayscale(Input, Output, Width, Height, Channels);
@@ -149,7 +149,7 @@ void LocalExponentialCorrection(unsigned char *Input, unsigned char *Output, int
     for (int mask = 0; mask < 256; mask++) {
         unsigned char *pLocalLut = LocalLut + (mask << 8);
         for (int pix = 0; pix < 256; pix++) {
-            pLocalLut[pix] = ClampToByte(255.0f * powf(pix / 255.0f, powf(2.0f, (128.0f - (255.0f - mask)) / 128.0f)));
+            pLocalLut[pix] = ClampToByte((int)(255.0f * powf((float)pix / 255.0f, powf(2.0f, (128.0f - (255.0f - mask)) / 128.0f))));
         }
     }
     Grayscale(Input, Luminance, Width, Height, Channels);
